requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-JWT==0.3.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==1.4.2
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
zipp==3.7.0
